rename 'get_caf_*' to 'get_cafs' + fix '_validate_allele' logic
