+++ src/anyvlm/restapi/vlm.py
@@ -20,7 +20,7 @@ from pydantic import BaseModel
 
 from anyvlm.anyvar.base_client import AnyVarClientConnectionError, BaseAnyVarClient
 from anyvlm.functions.build_vlm_response import build_vlm_response
-from anyvlm.functions.get_caf import VariantLookupError, get_caf
+from anyvlm.functions.get_cafs import VariantLookupError, get_cafs
 from anyvlm.functions.ingest_vcf import VcfAfColumnsError
 from anyvlm.functions.ingest_vcf import ingest_vcf as ingest_vcf_function
 from anyvlm.schemas.vlm import VlmResponse
@@ -313,7 +313,7 @@ def variant_counts(
     anyvlm_storage: Storage = request.app.state.anyvlm_storage
 
     try:
-        caf_data: list[AnyVlmCohortAlleleFrequencyResult] = get_caf(
+        caf_data: list[AnyVlmCohortAlleleFrequencyResult] = get_cafs(
             anyvar_client,
             anyvlm_storage,
             assemblyId,
