+++ src/anyvlm/storage/base_storage.py
@@ -39,7 +39,7 @@ class Storage(ABC):
         """
 
     @abstractmethod
-    def get_caf_by_vrs_allele_id(
+    def get_cafs_by_vrs_allele_id(
         self, vrs_allele_id: str
     ) -> list[AnyVlmCohortAlleleFrequencyResult]:
         """Retrieve cohort allele frequency study results by VRS Allele ID
