+++ tests/integration/functions/cassettes/test_get_caf/test_get_cafs_variant_not_registered.yaml
