+++ tests/integration/functions/test_get_cafs.py
@@ -1,11 +1,11 @@
-"""Test that get_caf function works correctly"""
+"""Test that get_cafs function works correctly"""
 
 import pytest
 from deepdiff import DeepDiff
 from helpers import EXPECTED_VRS_ID, TEST_VARIANT, build_caf
 
 from anyvlm.anyvar.python_client import PythonAnyVarClient
-from anyvlm.functions.get_caf import VariantLookupError, get_caf
+from anyvlm.functions.get_cafs import VariantLookupError, get_cafs
 from anyvlm.storage.postgres import PostgresObjectStore
 from anyvlm.utils.types import AnyVlmCohortAlleleFrequencyResult
 
@@ -20,13 +20,13 @@ def expected_cafs(caf_iri: AnyVlmCohortAlleleFrequencyResult, alleles: dict):
 
 
 @pytest.mark.vcr
-def test_get_caf_results_returned(
+def test_get_cafs_results_returned(
     anyvar_populated_python_client: PythonAnyVarClient,
     populated_postgres_storage: PostgresObjectStore,
     expected_cafs: list[AnyVlmCohortAlleleFrequencyResult],
 ):
-    """Test get_caf when variants are registered and results are expected"""
-    cafs = get_caf(
+    """Test get_cafs when variants are registered and results are expected"""
+    cafs = get_cafs(
         anyvar_populated_python_client,
         populated_postgres_storage,
         TEST_VARIANT.assembly,
@@ -44,12 +44,12 @@ def test_get_caf_results_returned(
 
 
 @pytest.mark.vcr
-def test_get_caf_no_results_returned(
+def test_get_cafs_no_results_returned(
     anyvar_populated_python_client: PythonAnyVarClient,
     postgres_storage: PostgresObjectStore,
 ):
-    """Test get_caf when variants are registered but no results are expected"""
-    cafs: list[AnyVlmCohortAlleleFrequencyResult] = get_caf(
+    """Test get_cafs when variants are registered but no results are expected"""
+    cafs: list[AnyVlmCohortAlleleFrequencyResult] = get_cafs(
         anyvar_populated_python_client,
         postgres_storage,
         TEST_VARIANT.assembly,
@@ -62,13 +62,13 @@ def test_get_caf_no_results_returned(
 
 
 @pytest.mark.vcr
-def test_get_caf_variant_not_registered(
+def test_get_cafs_variant_not_registered(
     anyvar_minimal_populated_python_client: PythonAnyVarClient,
     populated_postgres_storage: PostgresObjectStore,
 ):
-    """Test get_caf raises exception due to variant not being registered"""
+    """Test get_cafs raises exception due to variant not being registered"""
     with pytest.raises(VariantLookupError):
-        get_caf(
+        get_cafs(
             anyvar_client=anyvar_minimal_populated_python_client,
             anyvlm_storage=populated_postgres_storage,
             assembly_id=TEST_VARIANT.assembly,
